Refactor completo del codigo y arquitectura en clases devididas por funciones para evitar el archivo monolitico u hacer mas facil el mergeo. A futuro dar mas forma a las clases con entidades que tengan que ver. Me duele la cabeza en python y javascrupt esto es mas facil... por ahora te odio java ojala nos amiguemos nose aaaaaaaah................
