+++ BatallaNavalJava/nbproject/project.properties
@@ -37,7 +37,7 @@ includes=**
 jar.compress=false
 javac.classpath=
 # Space-separated list of extra javac options
-javac.compilerargs=
+javac.compilerargs=\ --enable-preview
 javac.deprecation=false
 javac.external.vm=true
 javac.modulepath=
@@ -82,7 +82,7 @@ run.classpath=\
 # Space-separated list of JVM arguments used when running the project.
 # You may also define separate properties like run-sys-prop.name=value instead of -Dname=value.
 # To set system properties for unit tests define test-sys-prop.name=value:
-run.jvmargs=
+run.jvmargs=\ --enable-preview
 run.modulepath=\
     ${javac.modulepath}
 run.test.classpath=\
