+++ BatallaNavalJava/src/batallanavaljava/Creditos.java
@@ -1,0 +1,15 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package batallanavaljava;
+
+/**
+ *
+ * @author Mkjdf983
+ */
+public class Creditos {
+        public static void creditos() {
+        System.out.println("Desarrollado por [Tu Nombre]");
+    }
+}
