+++ BatallaNavalJava/src/batallanavaljava/FinDelJuego.java
@@ -1,0 +1,16 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
+ */
+package batallanavaljava;
+
+/**
+ *
+ * @author Mkjdf983
+ */
+public class FinDelJuego {
+        public static void finDelJuego(String nombreJugador) {
+        System.out.println("Gracias por jugar, " + nombreJugador + "!");
+    }
+
+}
